Removed password and included a webdriver
Downloaded and included the geckodriver for Firefox. This is helpful when
trying to debug the script by following it along in the browser.

Installed dotenv to hold sensitive information. The README.md has been
updated to include instructions for using it as well as the webdrivers.

ghostdriver.log has been removed from the repo and .gitignore has been
instructed to not include any logs produced from the browsers and the .env file.

requirements.txt has been updated using pip freeze to include dotenv